Refactored TrackElement
